Fix JS and CSS combination minification
Closes #21
Closes #14

Basically we were not having the file re-evaluate and use the other compressors when its been put behind a front matter block. I added the change to both Page and Document as I was not sure which was the one absolutely needed but i did test this locally for myself and didnt run into any trouble. Perhaps you can make adjustments if you know something more than I do as I am still new to this.